+++ frontend/src/components/index.js
@@ -1,11 +1,13 @@
 export { default as ProfileChip } from './ProfileChip/ProfileChip';
 export { default as NavBar } from './NavBar/NavBar';
 export { default as DropdownMenu } from './DropdownMenu/DropdownMenu';
-export { default as Card, CARD_SIZE } from './Card/Card';
+export { default as Card } from './Card/Card';
+export { CARD_SIZE } from './Card/Card.styles';
 export { default as NewPostCard } from './Card/NewPostCard';
 export { default as EditPostCard } from './Card/EditPostCard';
 export { default as SelectBox } from './SelectBox/SelectBox';
-export { default as Button, BUTTON_SIZE } from './Button/Button';
+export { default as Button } from './Button/Button';
+export { BUTTON_SIZE } from './Button/Button.styles';
 export { default as PageButton } from './Button/PageButton';
 export { default as FilterList } from './FilterList/FilterList';
 export { default as CreatableSelectBox } from './CreatableSelectBox/CreatableSelectBox';
